refactor(Viewer): Move not primary component in /components folder
no code change. We only rename control.styl in vars.styl and move all the css in the folder
