+++ react/Viewer/components/InformationPanel.jsx
