+++ react/Viewer/components/PdfToolbarButton.jsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import Button from '../Button'
+
+import Button from '../../Button'
 
 const PdfToolbarButton = ({ icon, onClick, disabled, label }) => (
   <Button
