+++ react/Viewer/components/ViewerByFile.jsx
@@ -2,21 +2,21 @@ import React, { useMemo } from 'react'
 import PropTypes from 'prop-types'
 
 import { isMobile as isMobileDevice } from 'cozy-device-helper'
-import { FileDoctype } from '../proptypes'
-import withBreakpoints from '../helpers/withBreakpoints'
+import { FileDoctype } from '../../proptypes'
+import withBreakpoints from '../../helpers/withBreakpoints'
 
-import { isPlainText } from './helpers'
-import ImageViewer from './ViewersByFile/ImageViewer'
-import AudioViewer from './ViewersByFile/AudioViewer'
-import VideoViewer from './ViewersByFile/VideoViewer'
-import PdfJsViewer from './ViewersByFile/PdfJsViewer'
-import TextViewer from './ViewersByFile/TextViewer'
-import PdfMobileViewer from './ViewersByFile/PdfMobileViewer'
-import NoViewer from './NoViewer'
-import ShortcutViewer from './ViewersByFile/ShortcutViewer'
-import OnlyOfficeViewer from './ViewersByFile/OnlyOfficeViewer'
+import { isPlainText } from '../helpers'
+import ImageViewer from '../ViewersByFile/ImageViewer'
+import AudioViewer from '../ViewersByFile/AudioViewer'
+import VideoViewer from '../ViewersByFile/VideoViewer'
+import PdfJsViewer from '../ViewersByFile/PdfJsViewer'
+import TextViewer from '../ViewersByFile/TextViewer'
+import PdfMobileViewer from '../ViewersByFile/PdfMobileViewer'
+import NoViewer from '../NoViewer'
+import ShortcutViewer from '../ViewersByFile/ShortcutViewer'
+import OnlyOfficeViewer from '../ViewersByFile/OnlyOfficeViewer'
 
-import { useEncrypted } from './EncryptedProvider'
+import { useEncrypted } from '../EncryptedProvider'
 
 export const getViewerComponentName = ({
   file,
