+++ react/Viewer/components/styles.styl
@@ -1,19 +1,9 @@
+@require 'settings/z-index.styl'
 @require 'components/selectionbar.styl'
 @require 'settings/breakpoints.styl'
 @require 'settings/palette.styl'
 
-$toolbarHeight = 4rem
-$toolbarHeightMedium = 3rem
-$footerHeight = 3.5rem
-
-.viewer-controls
-    position relative
-    display flex
-    flex-direction column
-    justify-content center
-    align-items center
-    width 100%
-    height 100%
+@require '../vars.styl'
 
 .viewer-nav
     position absolute
@@ -50,12 +40,21 @@ $footerHeight = 3.5rem
     +medium-screen()
         display none
 
+.viewer-nav-arrow
+    opacity .7
+
+.viewer-controls
+    position relative
+    display flex
+    flex-direction column
+    justify-content center
+    align-items center
+    width 100%
+    height 100%
+
 .viewer-controls.--expanded .viewer-nav
     margin-top 0
     width 40%
-
-.viewer-nav-arrow
-    opacity .7
 
 .viewer-toolbar
     position absolute
