+++ react/Viewer/styles.styl
@@ -1,7 +1,7 @@
 @require 'settings/breakpoints.styl'
 @require 'settings/palette.styl'
 @require 'settings/z-index.styl'
-@require './controls.styl'
+@require './vars.styl'
 
 .viewer-wrapper
     position        absolute
