+++ react/Viewer/vars.styl
@@ -1,0 +1,3 @@
+$toolbarHeight = 4rem
+$toolbarHeightMedium = 3rem
+$footerHeight = 3.5rem
